refactor(utils): hoist polyak_update into shared RL utility
Move `polyak_update` out of bench-local `value_nets` into
`rlevo_reinforcement_learning::utils`, consolidating the
`ModuleVisitor`/`ModuleMapper` target-network logic that was
duplicated across every RL bench, integration test, and
bench-support module.

Remove now-unused imports (`HashMap`, `ModuleMapper`, `ModuleVisitor`,
`Param`, `ParamId`, `TensorData`) from all affected call sites and
switch them to the library import. Register all integration tests as
explicit `[[test]]` targets in the umbrella crate and add per-test
Just recipes, including `#[ignore]`-gated smoke runs for CI.

CRITICAL: `soft_update` in `ddpg_bench` now passes `&active.valid()`
(borrow) instead of `active.valid()` (moved value); verify the new
signature accepts a reference to avoid a double-borrow or moved-value
regression in downstream soft-update paths.
