+++ crates/rlevo/tests/sac_integration.rs
@@ -6,14 +6,14 @@
 //! reproducibility check are gated behind `#[ignore]` to stay inside the
 //! project's default-run integration-test budget.
 
-use std::collections::HashMap;
-
 use burn::backend::{Autodiff, Flex};
-use burn::module::{AutodiffModule, Module, ModuleMapper, ModuleVisitor, Param, ParamId};
+use burn::module::{AutodiffModule, Module};
 use burn::nn::{Linear, LinearConfig};
 use burn::tensor::activation::{relu, softplus, tanh};
 use burn::tensor::backend::{AutodiffBackend, Backend};
 use burn::tensor::{Tensor, TensorData};
+
+use rlevo_reinforcement_learning::utils::polyak_update;
 use rand::SeedableRng;
 use rand::rngs::StdRng;
 use serde::{Deserialize, Serialize};
@@ -324,46 +324,6 @@ impl<B: AutodiffBackend> ContinuousQ<B, 2, 2> for Critic<B> {
             tau as f32,
         )
     }
-}
-
-struct ParamCollector<B: Backend> {
-    tensors: HashMap<ParamId, TensorData>,
-    _marker: std::marker::PhantomData<B>,
-}
-impl<B: Backend> ModuleVisitor<B> for ParamCollector<B> {
-    fn visit_float<const D: usize>(&mut self, param: &Param<Tensor<B, D>>) {
-        self.tensors.insert(param.id, param.val().to_data());
-    }
-}
-struct PolyakMapper<B: Backend> {
-    active: HashMap<ParamId, TensorData>,
-    tau: f32,
-    _marker: std::marker::PhantomData<B>,
-}
-impl<B: Backend> ModuleMapper<B> for PolyakMapper<B> {
-    fn map_float<const D: usize>(&mut self, param: Param<Tensor<B, D>>) -> Param<Tensor<B, D>> {
-        let id = param.id;
-        let active = self.active.remove(&id).expect("paired active param");
-        let tau = self.tau;
-        param.map(move |t| {
-            let device = t.device();
-            let a = Tensor::<B, D>::from_data(active, &device);
-            t.mul_scalar(1.0 - tau) + a.mul_scalar(tau)
-        })
-    }
-}
-fn polyak_update<B: Backend, M: Module<B>>(active: &M, target: M, tau: f32) -> M {
-    let mut c = ParamCollector::<B> {
-        tensors: HashMap::new(),
-        _marker: std::marker::PhantomData,
-    };
-    active.visit(&mut c);
-    let mut m = PolyakMapper::<B> {
-        active: c.tensors,
-        tau,
-        _marker: std::marker::PhantomData,
-    };
-    target.map(&mut m)
 }
 
 type Be = Autodiff<Flex>;
